+++ src/L2Data2Code/L2Data2Code.SharedLib/Extensions/JSonExtensions.cs
@@ -17,5 +17,10 @@ namespace L2Data2Code.SharedLib.Extensions
             }
             return valueCollection;
         }
+
+        public static bool HasValues(this JArray array)
+        {
+            return array.HasValues;
+        }
     }
 }