+++ src/L2Data2Code/Mustache/IMustacheOptionsInitializer.cs
@@ -1,0 +1,9 @@
+using Newtonsoft.Json.Linq;
+
+namespace Mustache
+{
+    internal interface IMustacheOptionsInitializer
+    {
+        JToken Initialize(MustacheOptions options);
+    }
+}