+++ src/L2Data2Code/Mustache/MustacheAction.cs
@@ -1,6 +1,5 @@
 using L2Data2Code.BaseMustache.Interfaces;
 using L2Data2Code.SharedLib.Extensions;
-using L2Data2Code.SharedLib.Helpers;
 using Newtonsoft.Json.Linq;
 using System;
 using System.IO;
@@ -13,9 +12,10 @@ namespace Mustache
     internal class MustacheAction : IMustacheAction
     {
         private readonly IMustacheRenderizer renderizer;
-        private readonly IJsonSetting jsonSetting;
         private readonly IFileExecutor fileExecutor;
+        private readonly IMustacheOptionsInitializer mustacheOptionsInitializer;
         private MustacheOptions options;
+        private JToken entities;
 
         /// <summary>
         /// Constructor
@@ -24,11 +24,11 @@ namespace Mustache
         /// <param name="renderizer">Renderizer</param>
         /// <param name="fileExecutor">File executor</param>
         /// <exception cref="ArgumentNullException"></exception>
-        public MustacheAction(IJsonSetting jsonSetting, IMustacheRenderizer renderizer, IFileExecutor fileExecutor)
+        public MustacheAction(IMustacheRenderizer renderizer, IFileExecutor fileExecutor, IMustacheOptionsInitializer mustacheOptionsInitializer)
         {
-            this.jsonSetting = jsonSetting ?? throw new ArgumentNullException(nameof(jsonSetting));
             this.renderizer = renderizer ?? throw new ArgumentNullException(nameof(renderizer));
             this.fileExecutor = fileExecutor ?? throw new ArgumentNullException(nameof(fileExecutor));
+            this.mustacheOptionsInitializer = mustacheOptionsInitializer ?? throw new ArgumentNullException(nameof(mustacheOptionsInitializer));
         }
 
         /// <summary>
@@ -39,8 +39,8 @@ namespace Mustache
         public void Initialize(MustacheOptions options)
         {
             this.options = options ?? throw new ArgumentNullException(nameof(options));
-            jsonSetting.Initialize(options.JsonDataFile);
             fileExecutor.Initialize(options.TemplatePath);
+            entities = mustacheOptionsInitializer.Initialize(options);
         }
 
         /// <summary>
@@ -48,30 +48,27 @@ namespace Mustache
         /// </summary>
         public void Run()
         {
-            fileExecutor.RunOnPaths((path) => DoAction(CreatePath, path, options));
-            fileExecutor.RunOnFiles((file) => DoAction(TransformFile, file, options));
+            if (entities != null)
+            {
+                fileExecutor.RunOnPaths((path) => DoAction(CreatePath, path, options));
+                fileExecutor.RunOnFiles((file) => DoAction(TransformFile, file, options));
+            }
         }
 
         private void DoAction(Action<string, string, string, JToken> action, string path, MustacheOptions options)
         {
             var outputPath = options.OutputPath.AddPathSeparator();
             var sourcePath = options.TemplatePath.AddPathSeparator();
-            var view = jsonSetting.Config;
-            var entity = options.Collection;
-            var entities = string.IsNullOrWhiteSpace(entity) ? view : view.SelectToken(entity);
-
-            if (entities != null)
+            if (!Directory.Exists(outputPath))
             {
-                if (!Directory.Exists(outputPath))
-                {
-                    Directory.CreateDirectory(outputPath);
-                }
-
-                foreach (var item in entities)
-                {
-                    action.Invoke(path, outputPath, sourcePath, item);
-                }
+                Directory.CreateDirectory(outputPath);
             }
+
+            foreach (var item in entities)
+            {
+                action.Invoke(path, outputPath, sourcePath, item);
+            }
+
         }
 
         private void CreatePath(string path, string outputPath, string sourcePath, JToken item)
@@ -90,6 +87,5 @@ namespace Mustache
             File.WriteAllText(outpuFile, text);
             Console.WriteLine(outpuFile);
         }
-
     }
 }