+++ changes.md
@@ -1,3 +1,13 @@
+### 2.11.1
+
+- Upgrade NLog to 5.0.1 from 5.0.0ç
+- Added to MustacheHelpers *Join* with three variants:
+  - {{Join separator}}: Takes an array a join with a separator.
+  - {{JoinWithHeader separator header}}: Takes an array a join with a separator starting with a header.
+  - {{JoinWithHeaderFooter separator header footer}}: Takes an array a join with a separator starting with a header and ending with a footer.
+  - *separator, header and footer* are strings
+- Mustache add *IsFirst* and *IsLast* to nodes at JObject[] objects
+
 ### 2.11.0
 
 - Upgrade to Topelab.Core 1.4.0